Begin creating fetchPlayers function in Draft.jsx
This function will use an api to pull data from the .json files I have
created for the player cards. I have created the initial line, which
will use async for the api. The first section will check if isFetching
is running, and if it is, it will change isFetching to true and exit.