fetchPlayers logic in Draft.jsx completed
I have provisionally completed the fetchPlayers function within
Draft.jsx, prior to adding a catch error to the end of it. The function
will fetch the players and icons from the designated position groups
json file, it will parse them, then check for duplicates in the
selectedPlayers array (to prevent a player being selected twice). It
will then randomly select three players from the json file which will
be the available choices for the user to select.