Complete PlayerCard.jsx, amend index.css and tailwind.config
A provisionary build of the PlayerCard has now been completed, I have
added a test element within return() of Draft.jsx which was being used
to test different tailwind properties to construct the cards layout. Two
gradients created in tailwind css, one for standard cards, and one for
gold cards (ovr 80+) and this logic to display the relevant card is
on the PlayerCard.jsx file. I have tried to make them as responsive as
possible, cards should stay in a row when on a mobile screen, text
isnt always staying in line depending on section but I think this is
passable, for now at least.